Conflicts:
	Jenkinsfile

+++ scripts/start-long-running-instance.sh
@@ -2,7 +2,9 @@
 
 set -e
 
-rm -r grakn-new
+if [ -d grakn-new ]; then
+    rm -r grakn-new
+fi
 
 mkdir grakn-new
 
@@ -15,3 +17,7 @@ fi
 
 mv grakn-new grakn
 grakn/grakn server start
+
+grakn/graql console -k pokemon -f grakn/examples/pokemon.gql
+
+sudo systemctl restart periodic-query
